requirements.txt
# Telegram Bot
aiogram>=3.0.0

# Google APIs
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0

# Firebase/Firestore
google-cloud-firestore>=2.0.0

# Environment & Utilities
python-dotenv>=1.0.0
aiohttp>=3.0.0

# OpenAI (Whisper & Chat)
openai>=1.0.0
